requirements.txt
streamlit
PyMuPDF
fpdf
Pillow
PyPDF2
pdf2docx
pyttsx3
pytesseract
